lambda/requirements.txt
boto3
pinecone-client
pandas
